backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
pgvector-sqlalchemy

# ML/AI Models
torch
transformers[torch]
accelerate
ftfy
regex
git+https://github.com/openai/CLIP.git

# Agents & Tooling
openai
boto3
playwright
tqdm
python-dotenv